Added edits, comments, considerations, and todo to soil_profile.py
